requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
pandas
pydantic
pydantic-settings
alembic
python-dotenv
httpx